pyproject.toml
[project]
name = "rapha-pdf-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.117.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.37.0",
]
